requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
av==15.1.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.118.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
jiter==0.11.0
mpmath==1.3.0
numpy==2.3.3
onnxruntime==1.23.0
packaging==25.0
protobuf==6.32.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
setuptools==78.1.1
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
wheel==0.45.1
